Play updates
I modified the toggle for play to separate out the update play values so I can trigger the update outside of toggling edit/play. It is currently tied to  the threat level dropdown.

I changed the casing of a few values so I could use them directly in the interface.

I also fixed the calculations of attack XP so the appropriate sections have their numbers update as expected.