Skill update
I added a helper function for retrieving the inner html of an element.

I added variables for attribute mods.

I added retrieval of the competence bonus for skills that aren't signature skills; I also added retrieval of signature skill modifiers.

Skills are listed in the play section in a table, displaying the total, number of ranks, and default modifier. The total is the sum of ranks and modifiers. Ranks are calculated based on competence or signature skills.

For special characters, wounds are determined by Constitution score. Vitality is determined by threat level.

Size now also adjusts maximum wounds for special characters and damage saves for standard characters.

Defense is also now calculated based on size, Dexterity bonus, and threat level.